feat: add configuration for internationalization and utility to avoid use of no-null symbol and move environment file
