+++ l10n.yaml
@@ -1,0 +1,4 @@
+arb-dir: lib/l10n
+template-arb-file: en.arb
+output-localization-file: app_localizations.dart
+output-dir: lib/l10n/generated
